Fix parsing on 32bit systems
Explicitly use 64-bit values for parsing data to avoid errors on 32-bit
systems.

Fixes: https://github.com/prometheus-community/bind_exporter/issues/43

+++ bind/v2/v2.go
@@ -64,7 +64,7 @@ type Zone struct {
 
 type Counter struct {
 	Name    string `xml:"name"`
-	Counter uint   `xml:"counter"`
+	Counter uint64 `xml:"counter"`
 }
 
 // Client implements bind.Client and can be used to query a BIND v2 API.
